+++ examples/meg/topographic plotting.py
@@ -1,4 +1,6 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
+# This script can be executed from the Terminal using $ pythonw
+
 from eelbrain import *
 
 
@@ -8,32 +10,42 @@ w = 3
 ds = datasets.get_mne_sample(sub="modality=='A'", sns=True)
 
 
+# interactive plot
+plot.TopoButterfly('sns', ds=ds)
+
+
 # plot topography at a specific time point
-plot.Topomap("sns.sub(time=.07)", ds=ds, method='mne')
+plot.Topomap("sns.sub(time=.07)", ds=ds)
 
 # plot topography at a specific time point by condition
-plot.Topomap("sns.sub(time=.07)", 'side', ds=ds, method='mne')
+plot.Topomap("sns.sub(time=.07)", 'side', ds=ds)
 
-# plot topography at a specific time point for one condition
-plot.Topomap("sns.sub(side=='L', time=.07)", ds=ds, method='mne')
+# plot topography at a specific time point for one condition;
+# clip with a circular outline
+plot.Topomap("sns.sub(side=='L', time=.07)", ds=ds, clip='circular')
 
 # plot topography with a specific projection
-plot.Topomap("sns.sub(time=.07)", proj='back', ds=ds, method='mne')
+plot.Topomap("sns.sub(time=.07)", proj='back', ds=ds)
+
+# same but without clipping the map
+plot.Topomap("sns.sub(time=.07)", proj='back', clip=False, ds=ds)
+
 
 # mark sensors (all sensors with z coordinate < 0
 y = ds["sns"].sub(time=.07).mean('case')
 mark = ds['sns'].sensor.z < 0
-plot.Topomap(y, sensorlabels=False, mark=mark, method='mne', res=500)
+plot.Topomap(y, mark=mark)
 
-# non-circular projection, different cmap
-plot.Topomap(y, sensorlabels=False, interpolation='bilinear', res=500,
-             head_radius=(0.3, 0.33), head_pos=0.02, cmap='jet')
+# different cmap, and manually adjusted head outline
+plot.Topomap(y, head_radius=(0.3, 0.33), head_pos=0.02, cmap='jet')
 
-# sensor map alone, without labels
-plot.SensorMap(y, 'none', mark=mark, head_radius=(0.3, 0.33), head_pos=0.02)
+# sensor labels
+plot.SensorMap(y, mark=mark, head_radius=(0.3, 0.33), head_pos=0.02,
+               labels='name')
 
 # sensor map with labels
 p = plot.SensorMap(y, mark=mark, w=9, head_radius=(0.3, 0.33), head_pos=0.02)
+# move overlapping labels apart
 p.separate_labels()